Update day 3 solution.
A more elegant version of my initial approach. This could be made a lot more memory efficient by just storing the line segments and computing intersections. My idea would be to keep different sets for horizontal and vertical segments and than check the horizontals of the first path against the verticals of the second and vice versa.
Step computation than can use the sum of all segment length up to the intersecting segments and than adding the remaining steps of the intersecting segments.